% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cb_api.R
\name{check_cb_api}
\alias{check_cb_api}
\title{Check Coinbase API is working}
\usage{
check_cb_api(coinbase_url = "https://api.pro.coinbase.com/products")
}
\arguments{
\item{coinbase_url}{Url for Coinbase Feed API}
}
\value{
Response from Coinbase or an error message
}
\description{
Check Coinbase API is working
}
