% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{combine_trades}
\alias{combine_trades}
\title{Combine samples of trades}
\usage{
combine_trades(trades)
}
\arguments{
\item{trades}{Set of samples of tradea}
}
\value{
A data.frame of trades
}
\description{
Combine samples of trades
}
