% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{get_trades}
\alias{get_trades}
\title{Get recent trades}
\usage{
get_trades(coin_pair = "BTC-USD")
}
\arguments{
\item{coin_pair}{Currency pair: base - quote}
}
\value{
A data.frame of trades
}
\description{
Get recent trades from Coinbase
}
