% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.R
\name{get_trades_sample}
\alias{get_trades_sample}
\title{Get sample of recent trades}
\usage{
get_trades_sample(n = 10, coin_pair = "BTC-USD")
}
\arguments{
\item{n}{NUmber of samples}

\item{coin_pair}{Currency pair: base - quote}
}
\value{
A data.frame of trades
}
\description{
Get recent trades from Coinbase
}
